Backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["abhinav09 <shawabhinav93@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
django-cors-headers = "^4.6.0"
pillow = "^11.0.0"
model-bakery = "^1.20.0"
djangorestframework-simplejwt = "^5.3.1"
pyjwt = "^2.10.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
gunicorn = "^23.0.0"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
